Refactors the ProductService to improve code organization and exception handling.
Remove Retrieve Product before deleting in productService because this is already happen in repository.

Introduces core methods for product creation and updating to avoid code duplication.
fixing validation and exception handling for create, update, and delete operations.
Updates validation logic in DTOs to ensure ID values are greater than or equal to 1.
Adds unit tests for ProductService and validations using Moq and AutoFixture.