Updates order and return entities
Updates the order entity to use a string for the delivery company code and the return entity to use a GUID for the handled by user ID.

Adds a "Pending" status to the payment type enum.

Updates return statuses configuration with localized names and descriptions.

Configures the size entity to prevent client-side deletion of related product variants.